+++ NTUST-CSIE-EM-Project1/MyVector.h
@@ -9,26 +9,26 @@ class MyVector
 {
 public:
 	MyVector();
-	MyVector(std::string name, std::vector<double> data) :Name(name), Data(data) {}
+	MyVector(std::string name, std::vector<long double> data) :Name(name), Data(data) {}
 
 	//Operator overloading
 	MyVector operator+(const MyVector&);
 	MyVector operator-(const MyVector&);
 	MyVector operator*(const MyVector&);
 	MyVector operator/(const MyVector&);
-	double& operator[] (const int);
+	long double& operator[] (const int);
 
-	std::vector<double> GetData();
-	void SetData(std::vector<double>);
+	std::vector<long double> GetData();
+	void SetData(std::vector<long double>);
 	std::string GetName();
 	void SetName(std::string);
-	void InsertValue(double);
+	void InsertValue(long double);
 	
 
 private:
 	//向量名稱
 	std::string Name;
 	//向量資料
-	std::vector<double> Data;
+	std::vector<long double> Data;
 };
